requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
httpx==0.26.0
pandas==2.2.0
scikit-learn==1.4.0
streamlit==1.31.0
joblib==1.3.2
altair==5.2.0
watchdog==4.0.0
mkdocs-material==9.5.3
flake8==7.0.0
pytest==8.0.0
